agents/pyproject.toml
[tool.poetry]
name = "agents"
version = "0.1.0"
description = "Its an agent that is able to create CAD drawings"
authors = ["Gandhavalla Harshith <hgandavalla@gmail.com>"]
packages = [{include = "agents"}]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

[tool.poetry.group.agents.dependencies]
google-generativeai = "^0.8.4"
pyautogui = "^0.9.54"
python-dotenv = "^1.0.1"
opencv-python = "^4.11.0.86"
streamlit = "^1.42.0"
psutil = "^6.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
